+++ src/lib.rs
@@ -1,1 +1,1 @@
-pub mod lib;
+pub mod tools;
